backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
google-generativeai==0.3.2
python-docx==1.1.0
PyPDF2==3.0.1
textblob==0.19.0
nltk>=3.9
python-jose==3.3.0
python-decouple==3.8
aiofiles==23.2.0
motor==3.3.2
pymongo==4.6.1
passlib==1.7.4
bcrypt==4.1.2
email-validator==2.1.0
